--- src/Sat.java
@@ -1,6 +1,0 @@
-public class Sat {
-
-    public static void main(String[] args) {
-        // TODO
-    }
-}
